+++ src/decoding.rs
@@ -2,7 +2,7 @@ use base64::{engine::general_purpose::STANDARD, Engine};
 use serde::de::DeserializeOwned;
 
 use crate::algorithms::AlgorithmFamily;
-use crate::crypto::hmac::{HmacSecret, Hs256, Hs384, Hs512};
+use crate::crypto::hmac::HmacSecret;
 use crate::crypto::JwtVerifier;
 use crate::errors::{new_error, Error, ErrorKind, Result};
 use crate::header::Header;
@@ -12,6 +12,12 @@ use crate::pem::decoder::PemEncodedKey;
 use crate::serialization::{b64_decode, DecodedJwtPartClaims};
 use crate::validation::{validate, Validation};
 use crate::Algorithm;
+
+// Crypto
+#[cfg(feature = "aws_lc_rs")]
+use crate::crypto::aws_lc::hmac::{Hs256, Hs384, Hs512};
+#[cfg(feature = "rust_crypto")]
+use crate::crypto::rust_crypto::hmac::{Hs256, Hs384, Hs512};
 
 /// The return type of a successful call to [decode](fn.decode.html).
 #[derive(Debug)]
