+++ src/lib.rs
@@ -3,6 +3,11 @@
 //! Documentation:  [stable](https://docs.rs/jsonwebtoken/)
 #![deny(missing_docs)]
 
+#[cfg(all(feature = "rust_crypto", feature = "aws_lc_rs"))]
+compile_error!(
+    "feature \"rust_crypto\" and feature \"aws_lc_rs\" cannot be enabled at the same time"
+);
+
 pub use algorithms::Algorithm;
 pub use crypto::hmac::HmacSecret;
 pub use decoding::{decode, decode_header, DecodingKey, JwtDecoder, TokenData};
